[llgo]: fix compilation under current llvm
Patch rL322965 changed how intrinsics for memset and memzero
were defined. This causes a regression in LLGO making it no
longer buidable. In addition to that one pass was renamed,
so this also addresses the pass naming disparity.

I removed all split stack related bits from this patch as
as asked.

Differential Revision: https://reviews.llvm.org/D56638



git-svn-id: https://llvm.org/svn/llvm-project/llgo/trunk@351722 91177308-0d34-0410-b5e6-96231b3b80d8